requirements.txt
# ============================================================
# ARAG_PHARMA v3 — Dependencies
# ============================================================

# ── LLM: Groq (free tier, uses OpenAI SDK with custom base_url) ──
openai>=1.30.0
groq>=0.9.0

# ── Local Embeddings (FREE, runs on CPU) ─────────────────────
sentence-transformers>=2.7.0
torch>=2.0.0

# ── Vector Store (FREE, local FAISS) ─────────────────────────
faiss-cpu>=1.8.0

# ── API & Web ─────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0

# ── UI ────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Data & Utils ─────────────────────────────────────────────
pydantic>=2.7.0
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.7.0
lxml>=5.2.0

# ── Testing ───────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0

# ============================================================
# COST: $0.00/month for basic usage
#   Groq API free tier: 30 req/min, 14,400 req/day
#   DuckDuckGo web search: free, no key
#   Optional Serper: 2500 req/month free tier
# ============================================================
